km-bbb: copy System.map as well as part of installing kenrel
System.map isnt being copied currently and this patchset copies that file
as well to ensure that it copied to the install directory and then target
installation media as well.

This patchset also clears the old images built and so make sure to create
backups if required.

Signed-off-by: Krishna Konda <kkchaitu@gmail.com>

km-bbb: copy System.map as well as part of installing kenrel

System.map isnt being copied currently and this patchset copies that file
as well to ensure that it copied to the install directory and then target
installation media as well.

This patchset also clears the old images built and so make sure to create
backups if required.

Signed-off-by: Krishna Konda <kkchaitu@gmail.com>

km-bbb: fix some pathnames

Signed-off-by: Krishna Konda <kkchaitu@gmail.com>